pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fpl-predict"
version = "0.1.3"
description = "FPL player-points prediction & transfer advisor"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="FPL Predict"}]
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "pyarrow>=16.0",
    "requests>=2.31",
    "requests-cache>=1.2",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",          # ensure .env loading works
    "scikit-learn>=1.4",
    "xgboost>=2.0",                # gradient boosting
    "lightgbm>=4.0",               # gradient boosting alternative
    "optuna>=3.5",                 # hyperparameter optimization
    "matplotlib>=3.8",
    "pulp>=3.2",
    "click>=8.1",
    "PyYAML>=6.0",
    "tabulate>=0.9",
    "playwright>=1.45",            # browser automation for FPL auth
    "browser-cookie3>=0.19",
]

[project.scripts]
fpl = "fpl_predict.cli:main"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
strict_optional = true